move LinkView to ui package
