+++ code/app/src/main/res/layout/link_view.xml
@@ -3,7 +3,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent">
 
-    <com.example.experiment_automata.LinkView
+    <com.example.experiment_automata.ui.LinkView
         android:layout_width="wrap_content"
         android:layout_height="wrap_content" />
